Unidad 4/Sistema Deteccion de Emociones/requirements.txt
torch
torchvision
numpy
matplotlib
pillow
scikit-learn
opencv-python
